Half-hour session lengths (1, 1.5, 2, 2.5, 3)
Owner had to block off 2 hours because a 1.5-hour session couldn't be booked.
Sessions now come in half-hour steps everywhere: online, admin, and the reader.

Pricing: a half hour is the EXACT midpoint of the whole hours either side
(matrixPrice() interpolates), e.g. weekday 1 racer 1h $45 / 2h $85 -> 1.5h $65;
weekday 3 racers 1.5h = $187.50. Whole hours are unchanged. The extra-racer
add-on prorates too ($10/hr -> $15 for 1.5h).

Data (migration 014): bookings.duration_hours was an INTEGER, so 1.5 would have
been silently rounded. Now numeric(3,1), constrained to 1-3 in half-hour steps.
The discount hour caps (redemptions.hours, codes.hours_redeemed /
max_total_hours / max_hours_per_booking) are numeric as well, so a 1.5h
redemption no longer counts as 2h against a code's allowance.

Also fixed on the way:
- editBooking still rejected racer counts above 3 ("Racers must be 1, 2, or 3")
  — a leftover from the multi-racer change that would have blocked editing any
  big group. Now validates a whole number >= 1.
- PriceSummary computed the end time by adding whole hours, so a 1.5h session
  showed the wrong end. It now works in minutes.
- Reader DTOs typed durationHours as Int; with a numeric column Gson would
  THROW on 1.5 and take down the whole bookings list. Now Double, with an
  h-label helper so it renders "1.5h" not "1.5".

Reader v2.3 / versionCode 14: hour chips are now 1 / 1.5 / 2 / 2.5 / 3 in both
the walk-in session builder and Add-booking.

111 unit tests (6 new on the half-hour math and validator). Typecheck, web
build, and the signed Kotlin release build all pass.

+++ android-pos/app/src/main/java/com/mcracing/pos/net/Api.kt
@@ -23,7 +23,9 @@ data class BookingDto(
     val id: String,
     val sessionDate: String,
     val startTime: String,
-    val durationHours: Int,
+    // Half-hour sessions are supported, and the backend column is numeric — an
+    // Int here makes Gson throw on a 1.5 and takes down the whole list.
+    val durationHours: Double,
     val racerCount: Int,
     val sessionPriceCents: Long,
     // What staff should actually collect: session price minus any online discount.
@@ -130,7 +132,7 @@ data class CashPaymentRequest(
 data class CreateBookingRequest(
     val sessionDate: String,   // "YYYY-MM-DD"
     val startTime: String,     // "HH:MM" 24-hour
-    val durationHours: Int,
+    val durationHours: Double, // half-hour steps (1, 1.5, 2, 2.5, 3)
     val racerCount: Int,
     val priceCents: Long,
     val customerId: String? = null,
